backend/requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.28
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
